requirements.txt
pywin32
fastapi
uvicorn
pydantic
requests
groq
python-dotenv
watchdog